Correction code (doublons supprimés, code replacé)
